pyproject.toml
[project]
name = "reporter"
version = "0.1.0"
description = "Automate retrieval of Apple's reporter.jar reports"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dateutil>=2.9.0.post0",
    "structlog>=25.4.0",
    "typer>=0.16.0",
]

[project.scripts]
reporter = "reporter.cmd:main"

[build-system]
requires = ["uv_build>=0.8.5,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "mypy>=1.17.1",
    "pytest>=8.4.1",
    "ruff>=0.12.7",
]
